backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
aiosqlite>=0.20,<1.0
asyncpg>=0.29,<1.0
psutil>=5.9,<7.0
pyyaml>=6.0,<7.0
pydantic>=2.10,<3.0

